Server Deciding Alg Added
Added algorithm which decide which server to use from server list. 
On top of the job's parameters (core, memory and disk requirement), I've added loweestWaiting, which the server with  lowest number of waiting jobs. In other words, the least busy server.
I believe this is the best way to maximizes resource usage, which will bring down turnaround time.